% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tact-wrapper.R
\name{.tact_desc_tips}
\alias{.tact_desc_tips}
\title{Return descendant tip labels from a node}
\usage{
.tact_desc_tips(tree, node)
}
\arguments{
\item{tree}{An object of class \code{phylo}.}

\item{node}{Numeric node number.}
}
\value{
Character vector of descendant tip labels.
}
\description{
Return descendant tip labels from a node
}
\keyword{internal}
