% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tact-wrapper.R
\name{.tact_get_epithet}
\alias{.tact_get_epithet}
\title{Extract epithets from tip labels}
\usage{
.tact_get_epithet(labels)
}
\arguments{
\item{labels}{Character vector of labels.}
}
\value{
Character vector of everything after the first underscore, or \code{NA}.
}
\description{
Extract epithets from tip labels
}
\keyword{internal}
