% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tact-wrapper.R
\name{.tact_is_genus_only}
\alias{.tact_is_genus_only}
\title{Detect genus-only labels}
\usage{
.tact_is_genus_only(labels)
}
\arguments{
\item{labels}{Character vector of labels.}
}
\value{
Logical vector.
}
\description{
Detect genus-only labels
}
\keyword{internal}
