% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tact-wrapper.R
\name{.tact_make_unique}
\alias{.tact_make_unique}
\title{Make a label unique among used labels}
\usage{
.tact_make_unique(x, used)
}
\arguments{
\item{x}{Proposed label.}

\item{used}{Character vector of already-used labels.}
}
\value{
Unique label.
}
\description{
Make a label unique among used labels
}
\keyword{internal}
