% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tact-wrapper.R
\name{.tact_n_tips}
\alias{.tact_n_tips}
\title{Get tip counts for phylo or multiPhylo objects}
\usage{
.tact_n_tips(tree)
}
\arguments{
\item{tree}{A \code{phylo}, \code{multiPhylo}, or other object.}
}
\value{
Integer vector of tip counts, or \code{NA_integer_}.
}
\description{
Get tip counts for phylo or multiPhylo objects
}
\keyword{internal}
