% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tact-wrapper.R
\name{.tact_require}
\alias{.tact_require}
\title{Require R packages used by the TACT wrapper}
\usage{
.tact_require(pkgs = c("ape", "stringr", "phytools"))
}
\arguments{
\item{pkgs}{Character vector of package names.}
}
\value{
Invisibly returns \code{TRUE} if all packages are available.
}
\description{
Require R packages used by the TACT wrapper
}
\keyword{internal}
