% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tact-wrapper.R
\name{.tact_to_space}
\alias{.tact_to_space}
\title{Convert underscores to spaces}
\usage{
.tact_to_space(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector with all underscores converted to spaces.
}
\description{
Convert underscores to spaces
}
\keyword{internal}
