% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tact-wrapper.R
\name{.tact_to_underscore}
\alias{.tact_to_underscore}
\title{Convert spaces to underscores}
\usage{
.tact_to_underscore(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector with spaces converted to underscores.
}
\description{
Convert spaces to underscores
}
\keyword{internal}
