% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graft-primitives.R
\name{incoming_edge_index}
\alias{incoming_edge_index}
\title{Find the incoming edge of a node}
\usage{
incoming_edge_index(tr, node)
}
\arguments{
\item{tr}{An object of class \code{phylo}.}

\item{node}{Integer node number.}
}
\value{
The row index of the incoming edge, or \code{integer(0)} for the root.
}
\description{
Find the incoming edge of a node
}
\keyword{internal}
