% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tact-wrapper.R
\name{restore_tact_temp_names}
\alias{restore_tact_temp_names}
\title{Restore temporary TACT pseudo-genus names}
\usage{
restore_tact_temp_names(tree, map)
}
\arguments{
\item{tree}{An object of class \code{phylo} or \code{multiPhylo}.}

\item{map}{Mapping data frame returned by \code{\link[=split_nonmono_genera_for_tact]{split_nonmono_genera_for_tact()}}.}
}
\value{
A tree object of the same class with temporary prefixes removed.
}
\description{
Converts temporary labels such as \code{TACTTMPNeoponera01_apicalis} and
\code{TACTEXCLNeoponera01_bucki} back to \code{Neoponera_apicalis} and
\code{Neoponera_bucki}.
}
