% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_tree_paged}
\alias{plot_tree_paged}
\title{Plot a tree across multiple PDF pages}
\usage{
plot_tree_paged(
  tree,
  out_prefix = "tree_paged",
  tips_per_page = 2500,
  cex = 0.35,
  pdf_auto = TRUE,
  target_tpi = 16,
  min_h = 6,
  max_h = 200,
  buffer_in = 1.5,
  auto_width = TRUE,
  char_per_in_at_cex = 10,
  min_w = 5,
  max_w = 20,
  ladderize_it = TRUE,
  ...
)
}
\arguments{
\item{out_prefix}{Output prefix (each page becomes \verb{<prefix>_pageNN.pdf}).}

\item{tips_per_page}{Number of tips per page.}
}
\description{
Plot a tree across multiple PDF pages
}
